helper_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyserial==3.5
python-dotenv==1.0.1
httpx==0.27.2

